requirements.txt
psycopg2==2.9.9
python-dotenv==1.0.1
bs4==0.0.2
selenium==4.21.0
webdriver-manager==4.0.1
